+++ tests/qasm3_qir/converter/test_loop.py
@@ -18,6 +18,11 @@ import pytest
 
 from qbraid_qir.qasm3 import qasm3_to_qir
 from qbraid_qir.qasm3.visitor import Qasm3ConversionError
+from tests.qir_utils import (
+    check_attributes,
+    check_single_qubit_gate_op,
+    check_single_qubit_rotation_op,
+)
 
 EXAMPLE_WITHOUT_LOOP = """
 OPENQASM 3.0;
@@ -269,6 +274,115 @@ def test_convert_qasm3_for_loop_discrete_set():
     )
     assert str(qir_expected) == str(qir_from_loop)
     assert str(qir_from_loop) == EXAMPLE_QIR_OUTPUT
+
+
+def test_function_executed_in_loop():
+    """Test that a function executed in a loop is correctly parsed."""
+    qasm_str = """OPENQASM 3;
+    include "stdgates.inc";
+
+    def my_function(qubit q_arg, float[32] b) {
+        rx(b) q_arg;
+        return;
+    }
+    qubit[5] q;
+
+    int[32] n = 2;
+    float[32] b = 3.14;
+
+    for int i in [0:n] {
+        my_function(q[i], i*b);
+    }
+    """
+
+    result = qasm3_to_qir(qasm_str)
+    generated_qir = str(result).splitlines()
+
+    check_attributes(generated_qir, 5, 0)
+    check_single_qubit_rotation_op(generated_qir, 3, list(range(3)), [0, 3.14, 2 * 3.14], "rx")
+
+
+def test_loop_inside_function():
+    """Test that a function with a loop is correctly parsed."""
+    qasm_str = """OPENQASM 3;
+    include "stdgates.inc";
+
+    def my_function(qubit[3] q2) {
+        for int[32] i in [0:2] {
+            h q2[i];
+        }
+        return;
+    }
+    qubit[3] q1;
+    my_function(q1);
+    """
+
+    result = qasm3_to_qir(qasm_str)
+    generated_qir = str(result).splitlines()
+
+    check_attributes(generated_qir, 3, 0)
+    check_single_qubit_gate_op(generated_qir, 3, [0, 1, 2], "h")
+
+
+def test_function_in_nested_loop():
+    """Test that a function executed in a nested loop is correctly parsed."""
+    qasm_str = """OPENQASM 3;
+    include "stdgates.inc";
+
+    def my_function(qubit q_arg, float[32] b) {
+        rx(b) q_arg;
+        return;
+    }
+    qubit[5] q;
+
+    int[32] n = 2;
+    float[32] b = 3.14;
+
+    for int i in [0:n] {
+        for int j in [0:n] {
+            my_function(q[i], j*b);
+        }
+    }
+
+    my_function(q[0], 2*b);
+    """
+
+    result = qasm3_to_qir(qasm_str)
+    generated_qir = str(result).splitlines()
+
+    check_attributes(generated_qir, 5, 0)
+    check_single_qubit_rotation_op(
+        generated_qir,
+        9,
+        [0, 0, 0, 1, 1, 1, 2, 2, 2, 0],
+        [0, 3.14, 2 * 3.14, 0, 3.14, 2 * 3.14, 0, 3.14, 2 * 3.14, 2 * 3.14],
+        "rx",
+    )
+
+
+@pytest.mark.skip(reason="Not implemented nested functions yet")
+def test_loop_in_nested_function_call():
+    qasm3_string = """
+    OPENQASM 3;
+    include "stdgates.inc";
+    def my_function_1(qubit q1, int[32] a){
+        for int[32] i in [0:2]{
+            rx(a*i) q1;
+        }
+    }
+
+    def my_function_2(qubit q2, int[32] b){
+        my_function_1(q2, b);
+    }
+
+    qubit q;
+    my_function_2(q, 3);
+    """
+    result = qasm3_to_qir(qasm3_string)
+    generated_qir = str(result).splitlines()
+
+    check_attributes(generated_qir, 1, 0)
+    check_single_qubit_rotation_op(generated_qir, 3, [0, 0, 0], [0, 3, 6], "rx")
 
 
 def test_convert_qasm3_for_loop_unsupported_type():